APPSRE-5972 prepare fragment flow
